+++ packages/manager/src/components/TagsInput/TagsInput.tsx
@@ -15,7 +15,7 @@ export interface Tag {
   label: string;
 }
 
-export interface Props {
+export interface TagsInputProps {
   label?: string;
   hideLabel?: boolean;
   name?: string;
@@ -23,10 +23,10 @@ export interface Props {
   value: Item[];
   onChange: (selected: Item[]) => void;
   disabled?: boolean;
-  menuPlacement?: 'bottom' | 'top' | 'auto' | undefined;
+  menuPlacement?: 'bottom' | 'top' | 'auto';
 }
 
-const TagsInput: React.FC<Props> = (props) => {
+const TagsInput = (props: TagsInputProps) => {
   const {
     label,
     hideLabel,
@@ -113,4 +113,4 @@ const TagsInput: React.FC<Props> = (props) => {
   );
 };
 
-export default TagsInput;
+export { TagsInput };